Update
- Implementado função responsável por mostrar relatório de execução com a função 'Clock()'.

+++ trab3.c
@@ -9,6 +9,7 @@
 #include <sys/shm.h>
 #include <semaphore.h>
 #include <signal.h>
+#include <time.h>
 
 #define MEM_SZ 4096
 #define MEMO_Sema sizeof(int)
@@ -23,12 +24,13 @@ struct fila{
 	int sinal;
 	int totnum;
 	int StopAllProcess;
-//	int QtdP5;    
-//	int Qtdp6;
-	int turn;
-	int turn2;
- }; 
-int vet[11000];
+	int qtdp5;    
+	int qtdp6;
+	clock_t timebegin;
+ };
+ struct filaRel{
+	int dados[10000];	 
+ };
 // chanel for pipe
 int canal1[2],canal2[2]; 
 // responsible for create the fifo
@@ -40,7 +42,7 @@ void* transferePorPipe( void *ptr);
 // function resposible to transfer info of fifo1 whith pipe to process 6
  void* transferePorPipe2( void *ptr);
 // function resposible to send signal to p4
-void sinal(int p);
+void sinal();
 // clear fifo
 void clearfifo( struct fila *f );
 // vetor de numeros aleatórios
@@ -53,18 +55,23 @@ void* readp6( void *ptr );
 void* result( void *ptr );
 // response to creat shared memory for fifos
 void CreatSharedMemory(void ** shared_memory);
-void contadorf2( struct fila *f );
+void contadorf2( struct fila *f, int pipe);
 void changesinalf2( struct fila *f, int sinal);
 struct fila *fila_shared;
 struct fila *fila_shared2;
+struct filaRel *fila_shared3;
 void* result2( void *ptr );
 void* result3( void *ptr );
+void relp7();
 int main(){
+	clock_t timebegin, elapsedTime;
 	pid_t  pid, pid2, pid3, pid4,pid5, pid6, pid7; // processos 
 	pthread_t thread1,thread2,thread3,thread4,thread5,thread6,thread7[2];        // Threads
 	int i, nump5=0, nump6=0, cont,val;
 	void *shared_memory = (void *)0;
 	void *shared_memory2 = (void *)0;
+	void *shared_memory3 = (void *)0;
+	double finaltime;
 	//int shmid, shmid1; // shared memory id
 	
 	// transformar em função para f1 e f2
@@ -74,7 +81,8 @@ int main(){
 	CreatSharedMemory(&shared_memory2);
 	//printf("Memoria compartilhada no endereco=%p\n", shared_memory2);  
 	fila_shared2= (struct fila *) shared_memory2;
-	// vector with aleatori numbers 1-1000
+	fila_shared2->timebegin=clock();
+//	fila_shared3= (struct filaRel *) shared_memory3;
 	criarFila( fila_shared);
 	criarFila( fila_shared2);
 	
@@ -82,7 +90,7 @@ int main(){
 	if ( pipe(canal2) == -1 ){ printf("Erro pipe()"); return -1; } // canal 2 escrita fila2
     
     srand(5);
-	signal(SIGUSR1, sinal); // executar isso inicialmente no código
+	signal(SIGUSR1, sinal); 
 	for(int i=0;i<7;i++) // loop will run n times (n=5)
     {	if (i==0){
 			if(fork() == 0)
@@ -113,7 +121,6 @@ int main(){
 			if(fork() == 0)
 			{
 				// p4
-			//	printf("pid 2: %d\n",getpid());
 				pthread_create(&thread4, NULL, transferePorPipe2, NULL);   // inicia e executa o thread criado
 				transferePorPipe(NULL);	
 				pthread_join(thread4, NULL); // finaliza
@@ -125,16 +132,16 @@ int main(){
 			if(fork() == 0)
 			{
 				// p5
-				readp5(NULL);	
-				exit(0);
+		 	 	readp5(NULL);	
+				exit(0);   
 			}
 		}
 		else if(i==5){
 			if(fork() == 0)
 			{
 				// p6
-				readp6(NULL); 	
-				exit(0);
+			  	readp6(NULL); 	
+				exit(0);   
 			}
 
 		}
@@ -142,27 +149,17 @@ int main(){
 			if(fork() == 0)
 			{
 				// p7	
-			 	pthread_create(&thread7[0], NULL, result3, NULL);   // inicia e executa o thread criado
+  		  	  	pthread_create(&thread7[0], NULL, result3, NULL);   // inicia e executa o thread criado
 				pthread_create(&thread7[1], NULL, result2, NULL);   // inicia e executa o thread criado
 				result(NULL);
-				pthread_join(thread7[0], NULL); // finaliza 
-				pthread_join(thread7[1], NULL); // finaliza	 	 	
-				exit(0);
+				pthread_join(thread7[0], NULL); 
+				pthread_join(thread7[1], NULL); 	 
+				exit(0);    
 			}
 		}
     }
-
-	exit(0);
+	//exit(0);
 }
-	//exit(0); 
-void prenumale(int *vet){
-	int cont=0,num=0;
-	for(cont=0;cont<11000;cont++){
-		num=rand();
-		vet[cont]=(num=0? 1:num%1000);
-	}
-}
-
 // Function that creat a shared memory
 void CreatSharedMemory(void ** shared_memory)
 {	int shmid;
@@ -184,7 +181,7 @@ void CreatSharedMemory(void ** shared_memory)
 	}
 }
 
-void sinal(int p)
+void sinal()
 {    
 	fila_shared->sinal = 1;
 }
@@ -199,13 +196,10 @@ void* addfila ( void *ptr){
 					fila_shared->nItens++;
 					num=rand()%1000;
 					fila_shared->dados[fila_shared->nItens] = (num==0? 1:num);
-				} else { 	
-				//	fila_shared->sinal = 1;
-				//	while(fila_shared->sinal == 0); // change before
+				} 
+				if (fila_shared->nItens>8) { 	
 					kill(getpid(),SIGUSR1);
-				//	printf("entrei %d, pid: %d \n", fila_shared->sinal,getpid());
-				//	while(fila_shared->sinal == 1){};
-					}   // signal to p4
+				}   // signal to p4
 			}
 		sem_post((sem_t*)&fila_shared->mutex);
 	 }
@@ -225,10 +219,8 @@ void criarFila( struct fila *f) {
 	f->sinal = 0;
     f->totnum = 0;
 	f->StopAllProcess=1;
-//	f->QtdP5=0;
-//	f->Qtdp6=0;
-    f->turn=0;
-	f->turn2=0;
+	f->qtdp5=0;
+	f->qtdp6=0;
 
 }
 
@@ -244,10 +236,8 @@ void* transferePorPipe( void *ptr){
 	printf("pid 1: %d\n",getpid());
 	while (fila_shared2->StopAllProcess==1){
 		sem_wait((sem_t*)&fila_shared->mutex); // SEMA
-			if (fila_shared->sinal == 1 && fila_shared->nItens >-1){
+			if (fila_shared->sinal == 1 && fila_shared->nItens <9){
 				write(canal1[1],&fila_shared->dados[fila_shared->nItens],sizeof(int));			   // writing on chanel, conection of p4 and p5		
-		//		printf("p4 -> T1\n");fflush(stdout);
-				fila_shared->turn=1;
 				fila_shared->nItens--;
 				if(fila_shared->nItens==-1){
 					clearfifo(fila_shared )	;	 
@@ -261,13 +251,10 @@ void* transferePorPipe( void *ptr){
 }
 
 void* transferePorPipe2( void *ptr){
-	//printf("pid 2: %d\n",getpid());
 	while (fila_shared2->StopAllProcess==1){
 		sem_wait((sem_t*)&fila_shared->mutex); // SEMA			
-			if (fila_shared->sinal == 1 && fila_shared->nItens>-1){
+			if (fila_shared->sinal == 1 && fila_shared->nItens<9){
 				write(canal2[1],&fila_shared->dados[fila_shared->nItens],sizeof(int));			   // writing on chanel, conection of p4 and p6		
-			//	printf("p4 -> T2\n");fflush(stdout);
-				fila_shared->turn2=1;
 				fila_shared->nItens--; 
 				if(fila_shared->nItens==-1){
 					clearfifo(fila_shared); 
@@ -283,18 +270,15 @@ void* transferePorPipe2( void *ptr){
 void* readp5( void *ptr ){
 	int val, n=0;
 	while(fila_shared2->StopAllProcess==1){	
-	//	printf("aqui em baixo:");
-		while ( fila_shared2->sinal != 0 ){}
+		while ( fila_shared2->sinal != 0 ){if(fila_shared2->StopAllProcess==0)exit(0);}
 		if (fila_shared2->nItens<9) {
+			fila_shared2->qtdp5++;
 			fila_shared2->nItens++;
 			n=read(canal1[0],&fila_shared2->dados[fila_shared2->nItens],sizeof(int));
-	//		printf("pipe 1 -> p5 %d\n",n);fflush(stdout);
 			if (n==-1){
 				printf("Erro!!");
 				exit(0);
-				//	fila_shared2->QtdP5++;	
 			}
-			fila_shared->turn=0;
 		}
 		changesinalf2(fila_shared2,1);
 	}
@@ -305,33 +289,20 @@ void* readp5( void *ptr ){
 void* readp6( void *ptr ){
 	int val,n=0;
 	while(fila_shared2->StopAllProcess==1){	
-		while ( fila_shared2->sinal != 1 );
-		
-	//	if(fila_shared->turn2==1){
-			if (fila_shared2->nItens<9) {
-				fila_shared2->nItens++;
-				n=read(canal2[0],&fila_shared2->dados[fila_shared2->nItens],sizeof(int));
-	//			printf("pipe 2 -> p6 %d\n",n);fflush(stdout);
-				if (n==-1){
-					printf("Erro!!");
-					exit(0);
-					//	fila_shared2->QtdP5++;	
-				}
-				fila_shared->turn2=0;
+		while ( fila_shared2->sinal != 1 ){if(fila_shared2->StopAllProcess==0)exit(0);}
+		if (fila_shared2->nItens<9) {
+			fila_shared2->qtdp6++;
+			fila_shared2->nItens++;
+			n=read(canal2[0],&fila_shared2->dados[fila_shared2->nItens],sizeof(int));
+			if (n==-1){
+				printf("Erro!!");
+				exit(0);
 			}
-	//	}
+		}
 		changesinalf2(fila_shared2,2);
 	}
 	close(canal2[0]);
 	pthread_exit(0); /* exit thread */
-}
-
-void contadorf2( struct fila *f ){
-	f->nItens--; 
-	f->totnum++;
-	if (f->totnum==10000){
-		f->StopAllProcess=0;               // response to stop all process build 
-	}
 }
 
 void changesinalf2( struct fila *f, int sinal){
@@ -341,70 +312,71 @@ void changesinalf2( struct fila *f, int sinal){
 void* result( void *ptr ){
 	int cont=-1;
 	while (fila_shared2->StopAllProcess==1){
-	//	printf("t1 %d\n",fila_shared2->sinal);
-		while ( fila_shared2->sinal != 2 ){} // Busy wait
+		while ( fila_shared2->sinal != 2 ){if(fila_shared2->StopAllProcess==0)exit(0);} // Busy wait
 		if(fila_shared2->StopAllProcess==0)exit(0);
-		//printf("entrei aqui");
 		if(fila_shared2->nItens>-1){
-			printf("1 - numero: %d, %d\n",fila_shared2->dados[fila_shared2->nItens],fila_shared2->nItens);
+			printf("1 - numero: %d, %d\n",fila_shared2->dados[fila_shared2->nItens],fila_shared2->totnum);
 			fflush(stdout);
 			fila_shared2->nItens--; 
 			fila_shared2->totnum++;
 			if (fila_shared2->totnum==10000){
+				relp7();
 				fila_shared2->StopAllProcess=0;               // response to stop all process build 
 			}
 		}
 		changesinalf2(fila_shared2,3);	
 	}	
-	printf("deu certo   %d\n",fila_shared2->totnum);
-	//printf("deu certo  p5: %d, p6: %d\n",fila_shared2->QtdP5,fila_shared2->QtdP6);
-	// relatório:
+	//relp7();
 	pthread_exit(0); /* exit thread */
 }
 
 void* result2( void *ptr ){
 	int cont=-1;
 	while (fila_shared2->StopAllProcess==1){
-		while ( fila_shared2->sinal != 3 ){} // Busy wait
+		while ( fila_shared2->sinal != 3 ){if(fila_shared2->StopAllProcess==0)exit(0);} // Busy wait
 		if(fila_shared2->StopAllProcess==0)exit(0);
-		//printf("entrei aqui");
 		if(fila_shared2->nItens>-1){
-			printf("2 - numero: %d, %d\n",fila_shared2->dados[fila_shared2->nItens],fila_shared2->nItens);
+			printf("2 - numero: %d, %d\n",fila_shared2->dados[fila_shared2->nItens],fila_shared2->totnum);
 			fflush(stdout);
 			fila_shared2->nItens--; 
 			fila_shared2->totnum++;
 			if (fila_shared2->totnum==10000){
-				changesinalf2(fila_shared2,0);
+				relp7();
+				//changesinalf2(fila_shared2,0);
 				fila_shared2->StopAllProcess=0;               // response to stop all process build
 			}
 		}
 		changesinalf2(fila_shared2,4);	
 	}	
-	printf("deu certo   %d\n",fila_shared2->totnum);
-	//printf("deu certo  p5: %d, p6: %d\n",fila_shared2->QtdP5,fila_shared2->QtdP6);
-	// relatório:
+	//relp7();
 	pthread_exit(0); /* exit thread */
 }
 
 void* result3( void *ptr ){
 	int cont=-1;
 	while (fila_shared2->StopAllProcess==1){
-		while ( fila_shared2->sinal != 4 ){} // Busy wait
+		while ( fila_shared2->sinal != 4 ){if(fila_shared2->StopAllProcess==0)exit(0);} // Busy wait
 		if(fila_shared2->StopAllProcess==0)exit(0);
 		if(fila_shared2->nItens>-1){
-			printf("3 - numero: %d, %d\n",fila_shared2->dados[fila_shared2->nItens],fila_shared2->nItens);
+			printf("3 - numero: %d, %d\n",fila_shared2->dados[fila_shared2->nItens],fila_shared2->totnum);
 			fflush(stdout);
 			fila_shared2->nItens--; 
 			fila_shared2->totnum++;
 			if (fila_shared2->totnum==10000){
+				relp7();
 				fila_shared2->StopAllProcess=0;               // response to stop all process build 
 			}
 		}
 		changesinalf2(fila_shared2,0);	
 	}	
-	printf("deu certo   %d\n",fila_shared2->totnum);
-	//printf("deu certo  p5: %d, p6: %d\n",fila_shared2->QtdP5,fila_shared2->QtdP6);
-	// relatório:
+	//relp7();
 	pthread_exit(0); /* exit thread */
 }
 
+void relp7(){
+//	kill(getpid(),SIGSTOP);
+//	kill(getpid(),SIGCONT);
+	printf("--------------------------------------\n");fflush(stdout);
+	printf("Números impressos   %d \n",fila_shared2->totnum);fflush(stdout);
+	printf("Tempo %f  \n",(double)((clock()-fila_shared2->timebegin)/CLOCKS_PER_SEC));fflush(stdout);	
+}
